webcompstore-cli/requirements.txt
requests==2.25.1
urllib3==1.26.5
python_dotenv==0.19.2
python-keycloak==0.27.0